requirements.txt
Flask>=2.0
Werkzeug>=2.0
python-dotenv>=1.0
